refactor(QuizModule) : relocate ShortQuiz
- represent class as interface not DTO
