+++ src/quiz/interface/short-quiz.ts
@@ -1,8 +1,8 @@
-import { Artist } from '../../../artist/entities/artist.entity';
-import { Style } from '../../../style/entities/style.entity';
-import { Tag } from '../../../tag/entities/tag.entity';
-import { User } from '../../../user/entity/user.entity';
-import { Quiz } from '../../entities/quiz.entity';
+import { Artist } from '../../artist/entities/artist.entity';
+import { Style } from '../../style/entities/style.entity';
+import { Tag } from '../../tag/entities/tag.entity';
+import { User } from '../../user/entity/user.entity';
+import { Quiz } from '../entities/quiz.entity';
 
 export class ShortQuiz
   implements Pick<Quiz, 'title' | 'created_date' | 'id' | 'owner' | 'time_limit' | 'updated_date'>
