+++ src/quiz/quiz.controller.ts
@@ -45,7 +45,6 @@ import { CreateQuizDTO } from './dto/create-quiz.dto';
 import { GenerateQuizQueryDTO } from './dto/generate-quiz.query.dto';
 import { QuizResponseDTO } from './dto/output/response-quiz.dto';
 import { ResponseQuizDTO } from './dto/output/response-schedule-quiz.dto';
-import { ShortQuiz } from './dto/output/short-quiz.dto';
 import { QuizContextDTO } from './dto/quiz-context.dto';
 import { QuizReactionDTO, QuizReactionType } from './dto/quiz-reaction.dto';
 import { QuizReactionQueryDTO } from './dto/quiz-reaction.query.dto';
@@ -56,6 +55,7 @@ import { QuizDislike } from './entities/quiz-dislike.entity';
 import { QuizLike } from './entities/quiz-like.entity';
 import { Quiz } from './entities/quiz.entity';
 import { QuizContext } from './interface/quiz-context';
+import { ShortQuiz } from './interface/short-quiz';
 import { QuizScheduleService } from './quiz-schedule.service';
 import { QuizService } from './quiz.service';
 import { QuizCategory } from './type';
